main page 수정 -heojiyeon -19/12/05
