+++ WebContent/WEB-INF/views/search/searchHotel.jsp
@@ -1,7 +1,7 @@
 <%@ page language="java" contentType="text/html; charset=UTF-8"
     pageEncoding="UTF-8"%>
     
-<jsp:include page="/WEB-INF/views/layout/header.jsp"/>
+
 
 
 <style type="text/css">
@@ -12,10 +12,10 @@ width: 70%;
 
 
 </style>
+<h1>호텔 검색</h1>
 <div>
 <div  class="container text-center" id="hotel" data-skyscanner-widget="HotelSearchWidget" data-locale="ko-KR" 
 data-widget-scale="1.7"></div>
 <script src="https://widgets.skyscanner.net/widget-server/js/loader.js" async></script>
 </div>
 
-<jsp:include page="/WEB-INF/views/layout/footer.jsp"/>
