--- src/controller/login/MainController.java
@@ -1,25 +1,0 @@
-package controller.login;
-
-import java.io.IOException;
-
-import javax.servlet.ServletException;
-import javax.servlet.annotation.WebServlet;
-import javax.servlet.http.HttpServlet;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-
-/**
- * Servlet implementation class MainController
- */
-@WebServlet("/main")
-public class MainController extends HttpServlet {
-	private static final long serialVersionUID = 1L;
-
-	
-	@Override
-	protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-		req.getRequestDispatcher("/WEB-INF/views/main.jsp").forward(req, resp);
-		
-	}
-	
-}
